backend/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
click==8.1.8
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.12.2
uvicorn==0.34.0
cryptography~=44.0.2
PyJWT~=2.9.0
pytest~=8.3.5
pytest-json-report==1.5.0
pytest-cov==6.1.1
python-dotenv~=1.1.0
openai~=1.77.0
pytest-asyncio==0.26.0
aiosqlite>=0.18.0


jwt~=1.3.1
httpx~=0.28.1
aiosmtplib~=4.0.1